tests/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
duckdb==1.0.0
exceptiongroup==1.2.2
frozenlist==1.4.1
idna==3.7
iniconfig==2.0.0
multidict==6.0.5
packaging==24.1
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
requests==2.32.3
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
